requirements.txt
robotframework==7.0
robotframework-seleniumlibrary==6.2.0
selenium==4.21.0
webdriver-manager==4.0.1
faker==25.2.0
openai==0.28
